added label functionality (name above the box)
